refactor(tmux): consolidate tmux modules into src/tmux/ feature slice
- 11 files moved: 5 from src/tmux_*.rs, 5 from src/ui/tmux_*.rs, picker
- New src/tmux.rs aggregates TmuxSessionPlugin + 8 sub-plugins
- main.rs: 18->14 mod declarations, 8 singleton add_plugins removed
- Sub-plugin structs renamed (OzmuxTmux* prefix dropped, pub(crate))
- src/input/hyperlink.rs updated for new pane_hit path
