+++ src/tmux/copy_mode.rs
@@ -837,7 +837,7 @@ mod tests {
         // "scroll movement doesn't work" bug). The integration test below masked
         // this by starting a selection first.
         use crate::clipboard::Clipboard;
-        use super::render::RenderPlugin;
+        use crate::tmux::render::RenderPlugin;
         use bevy::window::{PrimaryWindow, Window, WindowResolution};
         use ozma_tty_renderer::material::TerminalUiMaterial;
         use ozma_tty_renderer::prelude::TerminalGridPlugin;
@@ -1018,7 +1018,7 @@ mod tests {
     #[ignore = "requires a real tmux binary and a controlling PTY"]
     fn copy_mode_integration_drives_real_tmux() {
         use crate::clipboard::Clipboard;
-        use super::render::RenderPlugin;
+        use crate::tmux::render::RenderPlugin;
         use bevy::window::{PrimaryWindow, Window, WindowResolution};
         use ozma_tty_renderer::material::TerminalUiMaterial;
         use ozma_tty_renderer::prelude::TerminalGridPlugin;
@@ -1244,7 +1244,7 @@ mod tests {
         // --- Assertions (all collected above so cleanup always runs first).
         assert!(
             snapshot_ok,
-            "OzmuxTmuxCopyModePlugin must receive at least one CopyModeSnapshot \
+            "CopyModePlugin must receive at least one CopyModeSnapshot \
              (state→capture round-trip) within 5 s of entering copy mode"
         );
         assert!(
