+++ src/tmux/mouse.rs
@@ -17,16 +17,16 @@
 //! on release. All copy-mode commands are pane-targeted (`send-keys -X -t %id`)
 //! so they act on the pressed pane regardless of the client's active pane.
 
+use super::copy_mode::{CopyModeSnapshot, cell_at_pane, cursor_deltas};
+use super::pane_hit::{cell_at_local, phys_to_pane_local, tmux_pane_at_phys};
+use super::render::{DividerPixelRect, PackedTmuxLayout};
 use crate::configs::OzmuxConfigsResource;
 use crate::inline_webview::{InlineWebview, inline_hit_at, inline_local_dip};
 use crate::input::InputPhase;
 use crate::input::current_modifiers;
 use crate::input::hyperlink::{link_modifier_held, should_open_at, try_open_uri};
 use crate::osc_webview::NonInteractive;
-use super::copy_mode::{CopyModeSnapshot, cell_at_pane, cursor_deltas};
-use super::pane_hit::{cell_at_local, phys_to_pane_local, tmux_pane_at_phys};
 use crate::picker::SessionPicker;
-use super::render::{DividerPixelRect, PackedTmuxLayout};
 use crate::ui::copy_mode::CopyModeState;
 use crate::ui::copy_search::CopyPrompt;
 use bevy::ecs::system::SystemParam;
